remove PoolMemberships provider, move types
