+++ src/hooks/useActiveBalances/index.tsx
@@ -23,7 +23,7 @@ import {
   defaultPayee,
 } from 'static/BalancesController/defaults';
 import type { PayeeConfig } from 'contexts/Setup/types';
-import type { PoolMembership } from 'contexts/Pools/PoolMemberships/types';
+import type { PoolMembership } from 'contexts/Pools/types';
 
 export const useActiveBalances = ({
   accounts,
@@ -105,7 +105,6 @@ export const useActiveBalances = ({
 
   // Gets an active balance's pool membership.
   const getPoolMembership = (address: MaybeAddress): PoolMembership | null => {
-    console.log('get pool membership from', activeBalances[address || '']);
     if (address) {
       const maybePoolMembership = activeBalances[address]?.poolMembership;
       if (maybePoolMembership) {
